s/mem_cpy/copy_from_slice/
