+++ src/decode.rs
@@ -1,0 +1,50 @@
+
+
+// Compute the evaluations of the error locator polynomial
+// `fn decode_init`
+// since this has only to be called once per reconstruction
+pub fn eval_error_polynomial(erasure: &[bool], log_walsh2: &mut [Multiplier], n: usize) {
+	let z = std::cmp::min(n, erasure.len());
+	for i in 0..z {
+		log_walsh2[i] = Multiplier(erasure[i] as Elt);
+	}
+	for i in z..n {
+		log_walsh2[i] = Multiplier(0);
+	}
+	walsh(log_walsh2, FIELD_SIZE);
+	for i in 0..n {
+		let tmp = log_walsh2[i].to_wide() * LOG_WALSH[i].to_wide();
+		log_walsh2[i] = Multiplier((tmp % ONEMASK as Wide) as Elt);
+	}
+	walsh(log_walsh2, FIELD_SIZE);
+	for i in 0..z {
+		if erasure[i] {
+			log_walsh2[i] = Multiplier(ONEMASK) - log_walsh2[i];
+		}
+	}
+}
+
+/// recover determines how many shards to recover (starting from 0)
+// technically we only need to recover
+// the first `k` instead of all `n` which
+// would include parity chunks.
+fn decode_main(codeword: &mut [Additive], recover_up_to: usize, erasure: &[bool], log_walsh2: &[Multiplier], n: usize) {
+	assert_eq!(codeword.len(), n);
+	assert!(n >= recover_up_to);
+	assert_eq!(erasure.len(), n);
+
+	for i in 0..n {
+		codeword[i] = if erasure[i] { Additive(0) } else { codeword[i].mul(log_walsh2[i]) };
+	}
+
+	inverse_afft(codeword, n, 0);
+
+	tweaked_formal_derivative(codeword, n);
+
+	afft(codeword, n, 0);
+
+	for i in 0..recover_up_to {
+		codeword[i] = if erasure[i] { codeword[i].mul(log_walsh2[i]) } else { Additive(0) };
+	}
+}
+
